backend/requirements.txt
asgiref
cachetools
certifi
charset-normalizer
Django==4.2
django-cors-headers
django-environ
djangorestframework
djangorestframework-simplejwt
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
httplib2
idna
protobuf
pyasn1
pyasn1-modules
PyJWT
pyparsing
pytz
requests
rsa
six
sqlparse
tzdata
uritemplate
urllib3
Pillow
psycopg2-binary